添加点赞功能

+++ life-base-api/src/main/java/com/dlion/life/base/entity/UserProjectRecord.java
@@ -16,9 +16,9 @@ public class UserProjectRecord {
 
     private Integer projectId;
 
-    private Byte isCreator;
+    private Integer isCreator;
 
-    private Byte attendStatus;
+    private Integer attendStatus;
 
     private Integer allPunchCardNum;
